requirements.txt
flask==2.3.3
torch==2.0.1
diffusers==0.25.0
transformers==4.35.2
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
spacy==3.6.1
scikit-learn==1.3.0
wikipedia-api==0.6.0
requests==2.31.0
python-dotenv==1.0.0
